move misc helpers to separate file
