+++ services/QuillLMS/client/app/bundles/Staff/components/evidence/__tests__/modelsTable.test.tsx
@@ -17,7 +17,7 @@ jest.mock("react-query", () => ({
   })),
 }));
 
-jest.mock('../../../helpers/evidence', () => ({
+jest.mock('../../../helpers/evidence/miscHelpers', () => ({
   titleCase: jest.fn().mockImplementation(() => {
     return '';
    }),
